refactor(charging-station): extract connector-status helpers to HelpersConnectorStatus (#1936 f-3)
Split Helpers.ts (Phase 3 of 5). Moves 8 connector-status symbols into
sibling HelpersConnectorStatus.ts and re-exports the public ones from
Helpers.ts so the barrel import path is preserved. Zero behavior change:
same signatures, same code, same log content. The moduleName local in
the new file switches log prefixes from 'Helpers.<fn>' to
'HelpersConnectorStatus.<fn>', reflecting the actual source location.

Moved (private inside new file):
- initializeConnectorStatus

Moved (exported from new file, re-exported from Helpers.ts):
- buildConnectorsMap, initializeConnectorsMapStatus, resetConnectorStatus,
  resetAuthorizeConnectorStatus, prepareConnectorStatus,
  checkStationInfoConnectorStatus, getBootConnectorStatus

Follows the barrel pattern established by HelpersReservation.ts in
Phase 1 (PR #1946).
